backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
google-genai
text2qti
pydantic
python-dotenv
